Move the new Logging back to the Monocle.Logging module
