+++ src/Macroscope/Main.hs
@@ -92,9 +92,9 @@ withMonitoringServer port logger action = do
 -- withClient "http://localhost:8080" Nothing $ runMacroscope 9001 "../etc/config.yaml"
 runMacroscope :: Int -> FilePath -> MonocleClient -> IO ()
 runMacroscope port confPath client = withLogger $ \logger -> withMonitoringServer port logger $ runLentilleM logger client $ do
-    logInfo_ "Starting to fetch streams"
-    config <- Config.mReloadConfig confPath
-    loop config (from ())
+  logInfo_ "Starting to fetch streams"
+  config <- Config.mReloadConfig confPath
+  loop config (from ())
  where
   loop config clients = do
     -- Load the config
