+++ src/Monocle/Prelude.hs
@@ -42,20 +42,6 @@ module Monocle.Prelude (
   Of (..),
   toVector,
 
-  -- * fast-logger
-  HasLogger (..),
-  runLogger,
-  runLogger',
-  logInfo,
-  logInfo_,
-  logWarn,
-  logWarn_,
-  logDebug,
-  logDebug_,
-  Logger,
-  LoggerT,
-  withLogger,
-
   -- * unliftio
   MonadUnliftIO,
   modifyMVar,
@@ -158,8 +144,6 @@ module Monocle.Prelude (
   BH.MonadBH,
   BH.DocId,
   BH.BulkOperation (..),
-  simpleSearch,
-  doSearch,
   mkAnd,
   mkOr,
   mkNot,
@@ -205,9 +189,7 @@ import Control.Monad.Writer (MonadWriter, WriterT, runWriterT, tell)
 import Data.Aeson (FromJSON (..), ToJSON (..), Value (Number, String), encode, object, withText, (.=))
 import Data.Aeson.Encode.Pretty qualified as Aeson
 import Data.Aeson.Key qualified as AesonKey
-import Data.Aeson.KeyMap qualified as KM
 import Data.Aeson.Lens (_Integer, _Object)
-import Data.Aeson.OneLine qualified
 import Data.Aeson.Types (Pair)
 import Data.Fixed (Deci, Fixed (..), HasResolution (resolution), Pico)
 import Data.Map qualified as Map
@@ -219,7 +201,6 @@ import Data.Vector (Vector)
 import Database.Bloodhound qualified as BH
 import GHC.Float (double2Float)
 import GHC.Generics (C, D, K1, M1, R, Rep, S, Selector, U1, selName, (:*:), (:+:))
-import GHC.Stack
 import Google.Protobuf.Timestamp qualified
 import Language.Haskell.TH.Quote (QuasiQuoter)
 import Network.HTTP.Client.OpenSSL (newOpenSSLManager, withOpenSSL)
@@ -235,7 +216,6 @@ import Streaming.Prelude (Stream)
 import Streaming.Prelude qualified as S
 import System.Environment (setEnv)
 import System.IO.Unsafe (unsafePerformIO)
-import System.Log.FastLogger qualified as FastLogger
 import Test.Tasty.HUnit
 import UnliftIO.Async (cancel, withAsync)
 import UnliftIO.MVar (modifyMVar, modifyMVar_)
@@ -364,78 +344,6 @@ dropMilliSec (UTCTime day sec) = UTCTime day (fromInteger $ truncate sec)
 headMaybe :: [a] -> Maybe a
 headMaybe xs = head <$> nonEmpty xs
 
------------------------------------------------------------
--- Logging facilities
-
-type Logger = FastLogger.TimedFastLogger
-
-class Monad m => HasLogger m where
-  getLogger :: m Logger
-  logIO :: IO () -> m ()
-
-newtype LoggerT a = LoggerT (ReaderT Logger IO a)
-  deriving newtype (Functor, Applicative, Monad, MonadIO, MonadReader Logger, MonadMask, MonadCatch, MonadThrow, MonadUnliftIO, Prometheus.MonadMonitor)
-
-runLogger :: MonadIO m => Logger -> LoggerT a -> m a
-runLogger logger (LoggerT action) = liftIO $ runReaderT action logger
-
-runLogger' :: MonadIO m => LoggerT a -> m a
-runLogger' (LoggerT action) = liftIO $ withLogger (runReaderT action)
-
-instance HasLogger LoggerT where
-  getLogger = ask
-  logIO = liftIO
-
-data LogLevel = LogWarning | LogInfo | LogDebug
-
-instance From LogLevel Text where
-  from = \case
-    LogWarning -> "WARNING "
-    LogInfo -> "INFO    "
-    LogDebug -> "DEBUG   "
-
-goLog :: HasLogger m => LogLevel -> Text -> Text -> [Pair] -> m ()
-goLog lvl loc msg attrs = do
-  logger <- getLogger
-  logIO $ logger (\time -> FastLogger.toLogStr $ time <> from msgText <> "\n")
- where
-  msgText :: Text
-  msgText =
-    from lvl <> loc <> ": " <> msg <> case attrs of
-      [] -> mempty
-      _ -> " " <> Data.Aeson.OneLine.renderObject (KM.fromList attrs)
-
-getLocName :: HasCallStack => Text
-getLocName = case getCallStack callStack of
-  (_ : (_, srcLoc) : _) -> from $ srcLocModule srcLoc <> ":" <> show (srcLocStartLine srcLoc)
-  _ -> "N/C"
-
-logInfo :: (HasCallStack, HasLogger m) => Text -> [Pair] -> m ()
-logInfo = goLog LogInfo getLocName
-
-logInfo_ :: (HasCallStack, HasLogger m) => Text -> m ()
-logInfo_ msg = goLog LogInfo getLocName msg []
-
-logWarn :: (HasCallStack, HasLogger m) => Text -> [Pair] -> m ()
-logWarn = goLog LogWarning getLocName
-
-logWarn_ :: (HasCallStack, HasLogger m) => Text -> m ()
-logWarn_ msg = goLog LogWarning getLocName msg []
-
-logDebug :: (HasCallStack, HasLogger m) => Text -> [Pair] -> m ()
-logDebug = goLog LogDebug getLocName
-
-logDebug_ :: (HasCallStack, HasLogger m) => Text -> m ()
-logDebug_ msg = goLog LogDebug getLocName msg []
-
--- | withLogger create the logger
-withLogger :: (Logger -> IO a) -> IO a
-withLogger cb = do
-  tc <- liftIO $ FastLogger.newTimeCache "%F %T "
-  FastLogger.withTimedFastLogger tc logger cb
- where
-  logger = FastLogger.LogStderr 1024
-
 getEnv' :: Text -> IO Text
 getEnv' var = do
   val <- from . exceptEnv <$> lookupEnv (from var)
@@ -554,24 +462,6 @@ toVector s = do
 
 -------------------------------------------------------------------------------
 -- Bloodhound helpers
-
--- | Helper search func that can be replaced by a scanSearch
-doSearch :: (HasLogger m, FromJSON a, MonadThrow m, BH.MonadBH m) => BH.IndexName -> BH.Search -> m (BH.SearchResult a)
-doSearch indexName search = do
-  -- logText . decodeUtf8 . encode $ search
-  rawResp <- BH.searchByIndex indexName search
-  -- logText $ show rawResp
-  resp <- BH.parseEsResponse rawResp
-  case resp of
-    Left e -> handleError e rawResp
-    Right x -> pure x
- where
-  handleError resp rawResp = do
-    logWarn "Elastic response failed" ["status" .= BH.errorStatus resp, "message" .= BH.errorMessage resp]
-    error $ "Elastic response failed: " <> show rawResp
-
-simpleSearch :: (FromJSON a, HasLogger m, MonadThrow m, BH.MonadBH m) => BH.IndexName -> BH.Search -> m [BH.Hit a]
-simpleSearch indexName search = BH.hits . BH.searchHits <$> doSearch indexName search
 
 mkAnd :: [BH.Query] -> BH.Query
 mkAnd andQ = BH.QueryBoolQuery $ BH.mkBoolQuery [] (BH.Filter <$> andQ) [] []
